fixing post date to number type
